Reactivated Diagram Synthesis
The current state of the plugin did not offer any way to activate.
The old way of opening a diagram from the context menu of some ecore file has been removed in the past.
This change introduces at least some rudimentary support to open the diagram alongside the ecore editor.

Signed-off-by: Nis Wechselberg <wechselberg.nisboerge@scheidt-bachmann-st.de>